+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 **This version includes Breaking Changes**
 - StatProperty#toString() no longer returns simple value text. Use StatProperty#valueText instead.
 - Renamed StatProperty#getFormattedValue() to getMultipliedValue().
+- Created DynamicTextAssets class for TextAssets which has placeholders.
+- Made SkillAttributeAssets extend DynamicTextAssets, which extends TextAssets.
+
 # 3.3.0
 - Added CharacterDetails#characterId.
 - Added CharacterDetails#getVoices().